requirements.txt
google-adk
litellm
pydantic
python-dotenv
streamlit
wikipedia
pythonmonkey